refactor: Remove worker type
